[#590] - Created a new route to show meals. Created a show meal page and redirect after saving from edit
